+++ button_handler.h
@@ -5,4 +5,4 @@
 #define STD_LOW 0
 
 void ButtonHandler();
-uint8_t Dio_readchannel(int channelID);
+uint8_t Dio_readchannel(uint8_t channelID);
